(#96) Metodos chart.js ficaram em app.js e exportados para popup.js
