+++ src/popup/popup.html
@@ -1,7 +1,7 @@
 <html>
 <head>     
     <script type='text/javascript' src='../CHART/chart.js'></script>
-    <script type='text/javascript' src='popup.js'></script>
+    <script type='module' src='popup.js'></script>
 </head>
 <body>
     <div class="container">
