+++ src/popup/popup.js
@@ -1,22 +1,9 @@
+import {GraficoPessoal} from '../CHART/app.js';
 document.addEventListener('DOMContentLoaded', function()
 {
-    var ctx = document.getElementById("chart").getContext('2d');
-    var myChart = new Chart(ctx, {
-        type: 'doughnut',
-        data: {
-            labels: [ 'a', 'b', 'c', 'd' ],
-                datasets: [{
-                backgroundColor: [
-                    "#59be5b",
-                    "#d56328",
-                    "#ff1b2d",
-                    "#0078d7"
-                ],
-                data: [ 1, 2, 3, 4 ]
-            }]
-        }
-    });
-
+   
+   
+    GraficoPessoal();
     document.getElementById('test').textContent = 'SUCCEED';
